backend/services/calendar_service.py
from datetime import datetime, timedelta
from google_auth_oauthlib.flow import Flow
from googleapiclient.discovery import build
from google.oauth2.credentials import Credentials
from config import GOOGLE_CLIENT_ID, GOOGLE_CLIENT_SECRET, GOOGLE_REDIRECT_URI

SCOPES = ["https://www.googleapis.com/auth/calendar.readonly"]


def get_auth_flow():
    """Create Google OAuth flow."""
    if not GOOGLE_CLIENT_ID or not GOOGLE_CLIENT_SECRET:
        return None

    client_config = {
        "web": {
            "client_id": GOOGLE_CLIENT_ID,
            "client_secret": GOOGLE_CLIENT_SECRET,
            "auth_uri": "https://accounts.google.com/o/oauth2/auth",
            "token_uri": "https://oauth2.googleapis.com/token",
            "redirect_uris": [GOOGLE_REDIRECT_URI],
        }
    }
    flow = Flow.from_client_config(client_config, scopes=SCOPES)
    flow.redirect_uri = GOOGLE_REDIRECT_URI
    return flow


def get_authorization_url():
    """Get the URL to redirect user to Google for OAuth consent."""
    flow = get_auth_flow()
    if not flow:
        return None
    url, _ = flow.authorization_url(
        access_type="offline",
        include_granted_scopes="true",
        prompt="consent",
    )
    return url


def exchange_code(code: str):
    """Exchange authorization code for tokens."""
    flow = get_auth_flow()
    if not flow:
        return None, None, None
    flow.fetch_token(code=code)
    credentials = flow.credentials
    return credentials.token, credentials.refresh_token, credentials.expiry


def get_upcoming_events(access_token, refresh_token, expires_at=None):
    """Fetch upcoming calendar events."""
    creds = Credentials(
        token=access_token,
        refresh_token=refresh_token,
        token_uri="https://oauth2.googleapis.com/token",
        client_id=GOOGLE_CLIENT_ID,
        client_secret=GOOGLE_CLIENT_SECRET,
    )

    service = build("calendar", "v3", credentials=creds)
    now = datetime.utcnow().isoformat() + "Z"
    week_later = (datetime.utcnow() + timedelta(days=7)).isoformat() + "Z"

    events_result = (
        service.events()
        .list(
            calendarId="primary",
            timeMin=now,
            timeMax=week_later,
            maxResults=20,
            singleEvents=True,
            orderBy="startTime",
        )
        .execute()
    )

    events = []
    for event in events_result.get("items", []):
        start = event["start"].get("dateTime", event["start"].get("date", ""))
        end = event["end"].get("dateTime", event["end"].get("date", ""))
        events.append({
            "id": event["id"],
            "summary": event.get("summary", "No title"),
            "start": start,
            "end": end,
            "location": event.get("location"),
        })
    return events, creds.token  # Return refreshed token
