backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
anthropic>=0.2.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-multipart>=0.0.5
httpx>=0.24.0
psycopg2-binary>=2.9.0
